Put openAndroidFile into FileUtils, tidy up
