+++ client/packages/config/src/config.ts
@@ -38,6 +38,7 @@ export const Environment = {
   PRINT_LABEL_QR: `${apiHost}/print/label-qr`,
   PRINT_LABEL_TEST: `${apiHost}/print/label-test`,
   PRINT_LABEL_PRESCRIPTION: `${apiHost}/print/label-prescription`,
+  ANDROID_DATA_FILES_PATH: `static_files/sync_files`,
 
   // -- Feature Flags --
   // To add a new feature flag:
